+++ src/hooks/useCardForm.js
@@ -1,131 +1,105 @@
-const useCardForm = () => {
-  const cardFormSchema = {
-    firstCardNumber: {
-      alias: '카드 번호',
-      type: 'number',
-      initialValue: '',
-      minLength: 4,
-      maxLength: 4,
-      required: true,
-    },
-    secondCardNumber: {
-      alias: '카드 번호',
-      type: 'number',
-      initialValue: '',
-      minLength: 4,
-      maxLength: 4,
-      required: true,
-    },
-    thirdCardNumber: {
-      alias: '카드 번호',
-      type: 'number',
-      initialValue: '',
-      minLength: 4,
-      maxLength: 4,
-      required: true,
-    },
-    fourthCardNumber: {
-      alias: '카드 번호',
-      type: 'number',
-      initialValue: '',
-      minLength: 4,
-      maxLength: 4,
-      required: true,
-    },
-    expiredMonth: {
-      alias: '만료일',
-      type: 'number',
-      initialValue: '',
-      minLength: 2,
-      maxLength: 2,
-      required: true,
-      validation: [
-        {
-          assert: (month) => month >= 1 && month <= 12,
-          errorMessage: '유효하지 않은 월입니다.',
-        },
-      ],
-    },
-    expiredYear: {
-      alias: '만료일',
-      type: 'number',
-      initialValue: '',
-      minLength: 2,
-      maxLength: 2,
-      required: true,
-      validation: [
-        {
-          assert: (year) => year >= new Date().getFullYear() % 100,
-          errorMessage: '유효하지 않은 연도입니다.',
-        },
-      ],
-    },
-    owner: {
-      alias: '카드 소유자 이름',
-      type: 'text',
-      initialValue: '',
-      maxLength: 30,
-    },
-    cvc: {
-      alias: '보안코드 (cvc/cvv)',
-      type: 'number',
-      initialValue: '',
-      minLength: 3,
-      maxLength: 4,
-      required: true,
-    },
-    firstPasswordDigit: {
-      alias: '비밀번호',
-      type: 'number',
-      initialValue: '',
-      minLength: 1,
-      maxLength: 1,
-      required: true,
-    },
-    secondPasswordDigit: {
-      alias: '비밀번호',
-      type: 'number',
-      initialValue: '',
-      minLength: 1,
-      maxLength: 1,
-      required: true,
-    },
-    thirdPasswordDigit: {
-      alias: '비밀번호',
-      type: 'number',
-      initialValue: '*',
-      minLength: 1,
-      maxLength: 1,
-    },
-    fourthPasswordDigit: {
-      alias: '비밀번호',
-      type: 'number',
-      initialValue: '*',
-      minLength: 1,
-      maxLength: 1,
-    },
+import { useState } from 'react';
+import useObjectRef from './useObjectRef';
+import useFormSchema from './useFormSchema';
+
+// formSchema를 useFormSchema를 이용하여 사용 가능한 형태로 분해
+// onChange, onKeyDown으로 에러 체크 및 포커싱 지원
+// 외부에서 정의한 onSubmit, onSubmitError 사용하여 handleSubmit 생성
+const useCardForm = ({ cardFormSchema, onSubmit, onSubmitError }) => {
+  const {
+    values,
+    setValues,
+    isInvalidInput,
+    errors,
+    setErrorTrue,
+    setErrorMessages,
+    focusNextElement,
+    focusPrevElement,
+  } = useFormSchema(cardFormSchema);
+
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const { ref, bindElement, getNextElement, getPrevElement } =
+    useObjectRef(cardFormSchema);
+
+  const handleChange = (event) => {
+    const {
+      target: { name, value },
+    } = event;
+
+    if (!Object.prototype.hasOwnProperty.call(values, name)) {
+      throw new Error('올바르지 않은 필드 참조입니다.');
+    }
+
+    if (isInvalidInput(name, value)) return;
+
+    setErrorMessages(name, value);
+
+    setValues((prevState) => ({
+      ...prevState,
+      [name]: value,
+    }));
+
+    focusNextElement(name, value, getNextElement(name));
   };
 
-  const addCard = async (values) => {
-    console.log(values);
+  const handleKeyDown = (event) => {
+    const {
+      target: { name },
+      keyCode,
+    } = event;
+    focusPrevElement(keyCode, name, getPrevElement(name));
   };
 
-  const onSubmit = async (values) => {
-    await addCard(values);
+  const handleSubmit = async (event) => {
+    event.preventDefault();
+
+    if (isSubmitting) {
+      return;
+    }
+
+    const invalidFieldNames = Object.keys(errors).filter(
+      (name) => errors[name].isError
+    );
+
+    if (invalidFieldNames.length > 0) {
+      const invalidInputRefs = invalidFieldNames
+        .map((name) => ref.current[name])
+        .sort((a, b) => a.id - b.id);
+
+      invalidFieldNames.forEach((name) => {
+        setErrorTrue(name);
+      });
+
+      onSubmitError(errors, invalidInputRefs);
+      return;
+    }
+
+    setIsSubmitting(true);
+    await onSubmit(values);
+    setIsSubmitting(false);
   };
 
-  const onSubmitError = (errors, invalidInputRefs) => {
-    const firstInvalidInput = invalidInputRefs[0].element;
-    const { errorMessage } = errors[firstInvalidInput.name];
-
-    firstInvalidInput.focus();
-    alert(errorMessage);
+  const registerInputProps = (name) => {
+    return {
+      name,
+      ref: bindElement(name),
+      value: values[name],
+      onChange: handleChange,
+      onKeyDown: handleKeyDown,
+    };
   };
+
+  const getInputClassName = (fieldName, additionalClassName = '') =>
+    `input-basic ${
+      errors[fieldName]?.showError ? 'error' : ''
+    } ${additionalClassName}`;
 
   return {
-    cardFormSchema,
-    onSubmit,
-    onSubmitError,
+    values,
+    isSubmitting,
+    handleSubmit,
+    registerInputProps,
+    getInputClassName,
   };
 };
